requirements.txt
fastapi>=0.111
uvicorn>=0.30
qdrant-client>=1.9
openai>=1.40
python-multipart>=0.0.9
pydantic>=2.7
sentence-transformers>=3.0
torch>=2.1,<3
pypdf>=4.2
python-dotenv>=1.0

